refactor: $props 관련 리팩토링
