+++ src/components/Todo/TodoListOfTeam.ts
@@ -1,7 +1,7 @@
 import {Component} from "@/core";
 import {TodoMemberAppender} from "./TodoMemberAppender";
 import {TodoList} from "./TodoList";
-import {todoOfTeamStore} from "@/store/todoOfTeamStore";
+import {todoOfTeamStore} from "@/store";
 
 export const TodoListOfTeam = class extends Component {
