+++ src/components/Todo/TodoMemberAppender.ts
@@ -1,5 +1,5 @@
 import {Component} from "@/core";
-import {SET_OPENED_APPEND_FORM, todoOfTeamStore} from "@/store/todoOfTeamStore";
+import {SET_OPENED_APPEND_FORM, todoOfTeamStore} from "@/store";
 
 export const TodoMemberAppender = class extends Component {
   template () {
